requirements.txt
requests
beautifulsoup4
moviepy
atproto
python-dotenv